IoT done

+++ A11/derek_gps/Iotracking_A11/WINC1500_HTTP_DOWNLOADER_EXAMPLE1/src/DistanceDriver/DistanceSensor.h
@@ -41,8 +41,8 @@
  ******************************************************************************/
 void InitializeDistanceSensor(void);
 void DeinitializeDistanceSerial(void);
-double parseGPRMC_lat(char stringtocheck[]);
-double parseGPRMC_long(char stringtocheck[]);
+float parseGPRMC_lat(char stringtocheck[]);
+float parseGPRMC_long(char stringtocheck[]);
 int32_t DistanceSensorGetDistance(float *distance, const TickType_t xMaxBlockTime); //Derek-@628
 void distUsartWritecallback(struct usart_module *const usart_module);
 void distUsartReadcallback(struct usart_module *const usart_module);
